Consolidate S3/local resource code, move to resource blueprint
There's a few things going on here:

1. Start moving resource code out of dataset blueprint
   and into a resource blueprint
2. Consolidate S3/local resource code
   Now that s3filestore returns a response object
   instead of immediately issueing a redirect:
   - The S3 resource_download can also call `log_download_activity()`
     only when we know we're going to issue the redirect
   - We can use the exact same code for both storage engines
   This is better because it is less code and more of it is under test
