+++ ckanext/unhcr/blueprints/__init__.py
@@ -6,6 +6,6 @@ from .dataset import unhcr_dataset_blueprint
 from .data_container import unhcr_data_container_blueprint
 from .deposited_dataset import unhcr_deposited_dataset_blueprint
 from .metrics import unhcr_metrics_blueprint
+from .resource import unhcr_resource_blueprint
 from .search_index import unhcr_search_index_blueprint
-from .storage import unhcr_s3_resource_blueprint
 from .user import unhcr_user_blueprint
